final-pjt-back/server/requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
enum34==1.1.10
idna==2.10
numpy==1.20.3
pandas==1.2.4
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rotten-tomatoes-client==0.0.3
rtsimple==0.9.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
